Add configuration service driver
